Add environment utilities and logging for AlphaFold setup
