+++ scripts/main.py
@@ -1,114 +1,59 @@
-import os
+import argparse
 import sys
-import subprocess
 from pathlib import Path
-from scripts.gpu_detect import detect_platform
+from scripts.env_utils import af_info, setup_platform
+from scripts.log_utils import check_required_file, log_message
 
 
-def log_message(message: str, log_path: Path) -> None:
-    """
-    Write a message to both stdout and append it to a log file.
-
-    Parameters:
-        message (str): The message to be logged.
-        log_path (Path): The file path where the log should be saved.
-    """
-    print(message)
-    with open(log_path, "a") as f:
-        f.write(message + "\n")
+class CustomArgParser(argparse.ArgumentParser):
+    def error(self, message):
+        self.print_help(sys.stderr)
+        self.exit(2, f"Error: {message}\n")
 
 
-def af_info(log_path: Path) -> None:
-    """
-    Collect and log Python environment information related to AlphaFold.
+def main():
+    parser = CustomArgParser(
+        description="Run AlphaFold workflow",
+        formatter_class=argparse.RawTextHelpFormatter
+    )
 
-    Parameters:
-        log_path (Path): Path to the log file where information will be written.
-    """
+    parser.add_argument(
+        "--input", 
+        type=Path, 
+        required=True,
+        help="Path to the input FASTA file"
+    )
 
-    try:
-        result = subprocess.run(
-            [sys.executable, "-m", "pip", "list"],
-            capture_output=True, 
-            text=True,
-            check=True
-        )
+    parser.add_argument(
+        "--mode",
+        type=str,
+        choices=["cpu", "gpu"], 
+        default="cpu", 
+        help="Mode to run: cpu or gpu, default: cpu"
+    )
 
-        jax_lines = [line for line in result.stdout.splitlines() if "jax" in line.lower()]
-        jax_versions = []
+    parser.add_argument(
+        "--log", 
+        type=Path, 
+        default=Path("run.log"), 
+        help="Path to the log file, default: ./run.log"
+    )
 
-        for line in jax_lines:
-            parts = line.split()
-            if len(parts) >= 2:
-                jax_versions.append(f"{parts[0]} {parts[1]}")
+    args = parser.parse_args()
 
-        jax_summary = ", ".join(jax_versions) if jax_versions else "N/A"
+    print("== AlphaFold 2 Runner ==")
+    print(f"Input file: {args.input}")
+    print(f"Run mode: {args.mode}")
+    print(f"Log file: {args.log}")
 
-    except FileNotFoundError:
-        log_message("Error: pip is not available in this environment.", log_path)
-        return
+    # Check if the input file exists
+    check_required_file(args.input)
 
-    tf_deterministic_ops = os.environ.get("TF_DETERMINISTIC_OPS", "N/A")
-    tf_unified_memory = os.environ.get("TF_FORCE_UNIFIED_MEMORY", "N/A")
-    xla_fraction = os.environ.get("XLA_PYTHON_CLIENT_MEM_FRACTION", "N/A")
+    # Setup the platform
+    setup_platform(args.mode)
 
-    log_message("==== AlphaFold Environment Information: ====", log_path)
-    log_message(f"Python version: {sys.version}", log_path)
-    log_message(f"JAX versions: {jax_summary}", log_path)
-    log_message(f"TF_DETERMINISTIC_OPS: {tf_deterministic_ops}", log_path)
-    log_message(f"TF_FORCE_UNIFIED_MEMORY: {tf_unified_memory}", log_path)
-    log_message(f"XLA_PYTHON_CLIENT_MEM_FRACTION: {xla_fraction}", log_path)
-    log_message(f"Log file path: {log_path}", log_path)
-    log_message("AlphaFold environment information collected successfully.", log_path)
-    log_message("==== End of Environment Information ====", log_path)
+    # Log the environment information
+    af_info(args.log)
 
-
-def setup_platform(mode: str) -> None:
-    """
-    Setup platform-specific environment variables for AlphaFold based on mode.
-
-    Parameters:
-        mode (str): Either "cpu" or "gpu". In "gpu" mode, the system auto-detects NVIDIA or AMD GPU.
-
-    Raises:
-        ValueError: If mode is unsupported.
-        RuntimeError: If GPU mode is selected but no platform is detected.
-    """
-    os.environ["TF_DETERMINISTIC_OPS"] = "1"
-
-    if mode == "cpu":
-        os.environ["JAX_PLATFORMS"] = "cpu"
-        # Try disabling both GPU type, just in case
-        os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
-        os.environ["HIP_VISIBLE_DEVICES"] = "-1"
-
-    elif mode == "gpu":
-        platform = detect_platform()
-        os.environ["XLA_PYTHON_CLIENT_MEM_FRACTION"] = "0.9"
-        
-        if platform == "nvidia":
-            os.environ["JAX_PLATFORMS"] = "cuda"
-            os.environ["CUDA_VISIBLE_DEVICES"] = "0"
-
-        elif platform == "amd":
-            os.environ["JAX_PLATFORMS"] = "rocm"
-            os.environ["HIP_VISIBLE_DEVICES"] = "0"
-            os.environ["ROCM_PATH"] = "/opt/rocm"
-    
-    else:
-        raise ValueError(f"Unsupported mode: {mode}")
-
-def check_required_file(file_path: Path) -> None:
-    """
-    Check whether the given file exists. If not, raise an error.
-
-    Parameters:
-        file_path (Path): The path to the file that should exist.
-
-    Raises:
-        FileNotFoundError: If the file does not exist.
-    """
-    if not file_path.exists():
-        raise FileNotFoundError(f"Missing required file: {file_path}")
-
-    print(f"Fasta file {file_path} exists.")
+    # Log the successful setup
+    print("Setup complete. Ready to run AlphaFold (not yet implemented).")
